subtle changes

+++ sysproc.c
@@ -106,24 +106,10 @@ sys_halt(void)
 int
 sys_signal(void)
 {
-  // cprintf("successfully calling the new syscall\n");
-  // cprintf("do the registration for the current process\n");
-  // proc->record = 1;
-
   int signum, handler;
   argint(0, &signum);
   argint(1, &handler);
-  // cprintf("signum = %d\n", signum);
-  // cprintf("handler = %d\n", handler);
-  // proc->signum = signum;
   proc->handler[signum] = handler;
-  // proc->fakeebp = proc->tf->ebp;
-  // proc->fakeesp = proc->tf->esp;
-  // uint ebp = proc->tf->ebp;
-  // uint esp = proc->tf->esp;
-  // ushort ss = proc->tf->ss;
-  // cprintf("sig ebp = %d\n", ebp);
-  // cprintf("sig esp = %d\n", esp);
-  // cprintf("sig ss = %d\n", ss);
-  return (int)(handler);
+
+  return handler;
 }
